- improving assertions system for commands and events - updating copyright
