--- .idea/artifacts/ExamplePluginToTest_jar.xml
@@ -1,8 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" build-on-make="true" name="ExamplePluginToTest:jar">
-    <output-path>$PROJECT_DIR$/../../MC/spigot_1.19.3/plugins</output-path>
-    <root id="archive" name="ExamplePluginToTest.jar">
-      <element id="module-output" name="ExamplePluginToTest" />
-    </root>
-  </artifact>
-</component>
