+++ src/main/java/io/github/jwdeveloper/spigot/tester/api/assertions/AssertionFactory.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c)  $originalComment.match("Copyright \(c\) (\d+)", 1, "-", "$today.year")2023. jwdeveloper
+ * Copyright (c)  2023  jwdeveloper
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,7 +24,8 @@
 
 package io.github.jwdeveloper.spigot.tester.api.assertions;
 
-import io.github.jwdeveloper.spigot.tester.api.EventCollector;
+import io.github.jwdeveloper.spigot.tester.api.collectors.CommandsCollector;
+import io.github.jwdeveloper.spigot.tester.api.collectors.EventCollector;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event;
 
@@ -33,9 +34,12 @@ import java.util.stream.Collectors;
 
 public class AssertionFactory {
     private final EventCollector eventCollector;
+    private final CommandsCollector commandsCollector;
 
-    public AssertionFactory(EventCollector collector) {
+    public AssertionFactory(EventCollector collector,
+                            CommandsCollector commandsCollector) {
         this.eventCollector = collector;
+        this.commandsCollector = commandsCollector;
     }
 
 
@@ -47,19 +51,17 @@ public class AssertionFactory {
         return new PlayerAssertions(player);
     }
 
-    public <T extends Event> EventsAssertions assertThatEvent(Class<T> eventClass) {
+    public <T extends Event> EventsAssertions<T> assertThatEvent(Class<T> eventClass) {
         var events = eventCollector
                 .getEvents()
                 .stream()
-                .filter(event ->
-                {
-                    return event.getClass().isAssignableFrom(eventClass);
-                })
+                .filter(event -> event.getClass().isAssignableFrom(eventClass) ||
+                        event.getClass().equals(eventClass))
                 .collect(Collectors.toList());
         return new EventsAssertions<T>((List<T>) events);
     }
 
     public CommandAssertion assertThatCommand(String command) {
-        return new CommandAssertion(command);
+        return new CommandAssertion(commandsCollector.getCommands(), command);
     }
 }
