+++ src/main/java/io/github/jwdeveloper/spigot/tester/api/assertions/CommandAssertion.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c)  $originalComment.match("Copyright \(c\) (\d+)", 1, "-", "$today.year")2023. jwdeveloper
+ * Copyright (c)  2023  jwdeveloper
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,48 +24,90 @@
 
 package io.github.jwdeveloper.spigot.tester.api.assertions;
 
+import io.github.jwdeveloper.spigot.tester.api.collectors.CommandInvokeResult;
+import org.bukkit.command.CommandSender;
+import org.bukkit.command.ConsoleCommandSender;
 import org.bukkit.entity.Player;
 
+import java.util.List;
+
 public class CommandAssertion {
-    private String command;
+    private final String command;
+    private final List<CommandInvokeResult> commands;
 
-    public CommandAssertion(String command) {
+    public CommandAssertion(List<CommandInvokeResult> commands, String command) {
         this.command = command;
+        this.commands = commands;
     }
 
-    public CommandAssertion wasInvoked(Times times)
-    {
+    public CommandAssertion wasInvoked(Times times) {
+        getAssertions(commands.size()).shouldBe(times.getValue());
         return this;
     }
 
-    public CommandAssertion byAnyone()
-    {
+    public CommandAssertion byPlayer() {
+        for (var event : commands) {
+            getAssertions(event.getCommandSender() instanceof Player).shouldBeTrue();
+        }
         return this;
     }
 
-    public CommandAssertion byPlayer(Player player)
-    {
+    public CommandAssertion byPlayer(Player player) {
+        for (var event : commands) {
+            getAssertions(event.getCommandSender() instanceof Player).shouldBeTrue();
+            getAssertions(event.getCommandSender()).shouldBe(player);
+        }
         return this;
     }
 
-    public CommandAssertion byConsole()
-    {
+    public CommandAssertion byCommandSender(CommandSender commandSender) {
+        for (var event : commands) {
+            getAssertions(event.getCommandSender()).shouldBe(commandSender);
+        }
         return this;
     }
 
-    public CommandAssertion withArguments(String ... args)
-    {
+    public CommandAssertion byConsole() {
+        for (var event : commands) {
+            getAssertions(event.getCommandSender() instanceof ConsoleCommandSender).shouldBeFalse();
+        }
+        return this;
+    }
+
+    public CommandAssertion withArguments(String... args) {
+
+        boolean found = false;
+        for (var cmd : commands) {
+            var i = 0;
+            for (var param : cmd.getCommandArgs()) {
+                if (!param.equals(args[i])) {
+                    break;
+                }
+                i++;
+            }
+            found = true;
+        }
+        getAssertions(found).shouldBeTrue();
         return this;
     }
 
     public CommandAssertion withSuccess()
     {
+        for (var event : commands) {
+            getAssertions(event.isInvokeResult()).shouldBe(true);
+        }
         return this;
     }
 
-    public boolean validate()
+    public CommandAssertion withFail()
     {
-        return false;
+        for (var event : commands) {
+            getAssertions(event.isInvokeResult()).shouldBe(false);
+        }
+        return this;
     }
 
+    private CommonAssertions getAssertions(Object target) {
+        return new CommonAssertions(target);
+    }
 }
