+++ src/main/java/io/github/jwdeveloper/spigot/tester/spigot/gson/JsonUtility.java
@@ -1,0 +1,61 @@
+/*
+ * MIT License
+ *
+ * Copyright (c)  2023  jwdeveloper
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package io.github.jwdeveloper.spigot.tester.spigot.gson;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import io.github.jwdeveloper.spigot.tester.spigot.gson.adapters.OffsetDateTimeAdapter;
+import io.github.jwdeveloper.spigot.tester.spigot.gson.adapters.ThrowAbleAdapter;
+
+import java.io.FileWriter;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.time.OffsetDateTime;
+
+public class JsonUtility {
+    public String save(Object data, String path, String fileName) throws IOException {
+        var outputPath = Paths.get(path, fileName + ".json");
+        Files.createDirectories(Paths.get(path));
+        try (FileWriter file = new FileWriter(outputPath.toString())) {
+            var gson = getGson();
+            var json = gson.toJson(data);
+            file.write(json);
+            return json;
+        } catch (IOException e) {
+            throw e;
+        }
+    }
+
+
+    private Gson getGson() {
+        return new GsonBuilder()
+                .registerTypeHierarchyAdapter(Throwable.class, new ThrowAbleAdapter())
+                .registerTypeHierarchyAdapter(OffsetDateTime.class, new OffsetDateTimeAdapter())
+                .setPrettyPrinting()
+                .create();
+    }
+
+}
