+++ src/main/java/io/github/jwdeveloper/spigot/tester/spigot/gson/adapters/OffsetDateTimeAdapter.java
@@ -1,0 +1,56 @@
+/*
+ * MIT License
+ *
+ * Copyright (c)  2023  jwdeveloper
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package io.github.jwdeveloper.spigot.tester.spigot.gson.adapters;
+
+import com.google.gson.*;
+
+import java.lang.reflect.Type;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeFormatterBuilder;
+import java.time.temporal.ChronoField;
+
+
+public class OffsetDateTimeAdapter implements JsonSerializer<java.time.OffsetDateTime>, JsonDeserializer<java.time.OffsetDateTime> {
+
+    private final DateTimeFormatter formatter;
+
+    public OffsetDateTimeAdapter() {
+        formatter = new DateTimeFormatterBuilder()
+                .appendPattern("yyyy-MM-dd HH:mm:ss")
+                .optionalStart().appendFraction(ChronoField.NANO_OF_SECOND, 0, 9, true).optionalEnd()
+                .appendPattern("x")
+                .toFormatter();
+    }
+
+    @Override
+    public java.time.OffsetDateTime deserialize(JsonElement json, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+        return java.time.OffsetDateTime.parse(json.getAsString(), this.formatter);
+    }
+
+    @Override
+    public JsonElement serialize(java.time.OffsetDateTime offsetDateTime, Type type, JsonSerializationContext jsonSerializationContext) {
+        return new JsonPrimitive(offsetDateTime.format(this.formatter));
+    }
+}
